Appear to have fixed error where holding screen sideways resulted in different views overlapping or the screen going black.
The problem appears stem from the fact the application uses a "launcher" activity to decide which activity to go to.  Because that launcher activity was not a "portrait" activity, the system got confused when an intent was sent to that activity, it was initiated with a change in screen orientation.